% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U_series_gain_loss_function.R
\name{U_series_gain_loss_function}
\alias{U_series_gain_loss_function}
\title{U_series_gain_loss_function}
\usage{
U_series_gain_loss_function(
  input_data,
  nbit = 1000,
  logT_min = 1,
  logT_max = 7,
  Tmult_min = 0.5,
  logk238_min = -7,
  logk238_max = -5,
  k48_min = 1,
  k48_max = 2,
  logf238_k238_min = -8,
  logf238_k238_max = +0.5,
  f48_min = 0.1,
  f48_max = 10
)
}
\arguments{
\item{input_data}{Read a CSV file with the required data. For instance, if the CSV file is called "data.csv" and is in your working directory, have input_data = read.csv("data.csv")}

\item{nbit}{Number of iterations}

\item{logT_min}{Log10 of the minimum value allowed for the weathering age of the shallowest sample}

\item{logT_max}{Log10 of the maximum value allowed for the weathering age of the shallowest sample}

\item{Tmult_min}{Minimum value allowed for Tmult}

\item{logk238_min}{Log10 of the minimum value allowed for k238}

\item{logk238_max}{Log10 of the maximum value allowed for k238}

\item{k48_min}{Minimum value allowed for k234/k238}

\item{k48_max}{Maximum value allowed for k234/k238}

\item{logf238_k238_min}{Log10 of the minimum value allowed for f238/k238}

\item{logf238_k238_max}{Log10 of the maximum value allowed for f238/k238}

\item{f48_min}{Minimum value allowed for f234/f238}

\item{f48_max}{Maximum value allowed for f234/f238}
}
\value{

}
\description{
U_series_gain_loss_function
}
\examples{
U_series_gain_loss_function (read.csv(system.file("extdata", "data.csv",
package = "UseriesGAwigl")),
logT_min = 1, logT_max = 7, Tmult_min = 0.5,
logk238_min = -7, logk238_max = -5,
k48_min = 0.1, k48_max = 10,
logf238_k238_min = -20, logf238_k238_max = 3,
f48_min = 0.1, f48_max = 10)
}
